requirements.txt
Flask
mysql-connector-python
fuzzywuzzy
python-Levenshtein
gTTS
python-dotenv
gunicorn
uwsgi==2.0.30